Use Philox64x4 PRNG, and removal RETURN_ENCODED_SOFTMAX=True variant from compiled forward kernel (#71)
## Major Changes

* [kernel] Use 64-bit Philox PRNG instead of 32-bit, to support 64-bit offset input
    + This also uses all four 64-bit outputs of PRNG instead of only keeping one result
* [kernel] add new kernel `debug_simulate_encoded_softmax`
    + [shim] Expose kernel `debug_simulate_encoded_softmax`
* [shim] Let `attn_fwd` call `debug_simulate_encoded_softmax` when encoded softmax is needed
    + `RETURN_ENCODED_SOFTMAX=True` option will not be used in the new `attn_fwd`
* [shim] Do not compile `RETURN_ENCODED_SOFTMAX=True` attn_fwd kernel anymore.

## Minor Changes

* [build] Use `pip install .` instead of `python setup.py develop` to install Triton in `build/venv`
    + This is less error prone
* Pass null tensors to philox arguments if `dropout_p <= 0.0`
    + This is suggested by the performance downgrade detected in main_perf kernel.

## Known Problems

* `debug_simulate_encoded_softmax` returns all 0.5/-0.5 instead of actual QK.
  In other words, only the sign bit is meaningful. This is by design since we
  will not add QK inner product in this function, and users of encoded softmax
  do not care QK as well.
* Old tuning database does not always work due to increased register pressures
    + Faulty kernels generate NaN due to spill registers. 
* attn_fwd tuning database is still missing

+++ v2src/flash/attn_bwd.cc
@@ -1,4 +1,4 @@
-// Copyright © 2023-2024 Advanced Micro Devices, Inc.
+// Copyright © 2023-2025 Advanced Micro Devices, Inc.
 // SPDX-License-Identifier: MIT
 
 #include <aotriton/config.h>
@@ -207,7 +207,7 @@ bwd_kernel_dk_dv(T4 q,
     .dropout_p = dropout_p,
     .philox_seed_ptr = &philox_seed,
     .philox_offset1 = &philox_offset1,
-    .philox_offset2 = static_cast<uint32_t>(philox_offset2),
+    .philox_offset2 = static_cast<uint64_t>(philox_offset2),
     .BLOCK_DMODEL = head_size_rounded,
     .CAUSAL = is_causal,
     .ENABLE_DROPOUT = dropout_p > 0.0,
@@ -322,7 +322,7 @@ bwd_kernel_dq(T4 q,
     .dropout_p = dropout_p,
     .philox_seed_ptr = &philox_seed,
     .philox_offset1 = &philox_offset1,
-    .philox_offset2 = static_cast<uint32_t>(philox_offset2),
+    .philox_offset2 = static_cast<uint64_t>(philox_offset2),
     .BLOCK_DMODEL = head_size_rounded,
     .CAUSAL = is_causal,
     .ENABLE_DROPOUT = dropout_p > 0.0,
